+++ nctest/vargetg.c
@@ -64,7 +64,7 @@ test_ncvargetg(path)
 	}
 
 	/* get space for vals */
-	hc.vals = emalloc(nctypelen(test.vars[iv].type) + 8);
+	hc.vals = emalloc((size_t)nctypelen(test.vars[iv].type) + 8);
 
 #	define TEST_FAILS(varid) \
 	    (ncvarputg(cdfid, varid, hc.cor, hc.edg, \
